Move node registry validation functions up
Moves the node registry validation functions up to the top level node
registry module, since they will be used throughout the module.
