+++ libsplinter/src/node_registry/mod.rs
@@ -247,3 +247,47 @@ where
         (**self).delete_node(identity)
     }
 }
+
+/// Returns `Err` if not all `nodes` are valid.
+fn validate_nodes(nodes: &[Node]) -> Result<(), InvalidNodeError> {
+    for (idx, node) in nodes.iter().enumerate() {
+        check_node_required_fields_are_not_empty(node)?;
+        check_if_node_is_duplicate(node, &nodes[idx + 1..])?;
+    }
+    Ok(())
+}
+
+/// Checks emptiness properties of all fields on the given `node`.
+fn check_node_required_fields_are_not_empty(node: &Node) -> Result<(), InvalidNodeError> {
+    if node.identity.is_empty() {
+        Err(InvalidNodeError::EmptyIdentity)
+    } else if node.endpoints.is_empty() {
+        Err(InvalidNodeError::MissingEndpoints)
+    } else if node.endpoints.iter().any(|endpoint| endpoint.is_empty()) {
+        Err(InvalidNodeError::EmptyEndpoint)
+    } else if node.display_name.is_empty() {
+        Err(InvalidNodeError::EmptyDisplayName)
+    } else {
+        Ok(())
+    }
+}
+
+/// Checks if the given `node` is a duplicate of any in the slice of `existing_nodes`.
+fn check_if_node_is_duplicate(
+    node: &Node,
+    existing_nodes: &[Node],
+) -> Result<(), InvalidNodeError> {
+    existing_nodes.iter().try_for_each(|existing_node| {
+        if existing_node.identity == node.identity {
+            Err(InvalidNodeError::DuplicateIdentity(node.identity.clone()))
+        } else if let Some(endpoint) = existing_node
+            .endpoints
+            .iter()
+            .find(|endpoint| node.endpoints.contains(endpoint))
+        {
+            Err(InvalidNodeError::DuplicateEndpoint(endpoint.clone()))
+        } else {
+            Ok(())
+        }
+    })
+}
